CRM學生快速打卡/check-in-crm/pyproject.toml
[project]
name = "check-in-crm"
version = "0.1.0"
description = "Fast Event Check-In CRM for collecting emails at Chinese language teaching department events"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.47.0",
    "google-auth-oauthlib>=1.2.2",
    "jinja2>=3.1.6",
    "prisma>=0.15.0",
    "python-multipart>=0.0.21",
    "uvicorn[standard]>=0.40.0",
]

[project.scripts]
dev = "uvicorn main:app --reload"
